// projects/evo-ui-kit/src/lib/common/enums/evo-color.ts
export enum EvoColor {
    primary = 'primary',
    secondary = 'secondary',
    success = 'success',
    secondaryLight = 'secondary-light',
    bonus = 'bonus',
    backgroundGrey = 'background-grey',
    highlight = 'highlight',
    rating = 'rating',
    primaryLight = 'primary-light',
    iconLight = 'icon-light',
    iconDark = 'icon-dark',
    disabled = 'disabled',
    text = 'text',
    textGrey = 'text-grey',
    textSubscription = 'text-subscription',
    link = 'link',
    error = 'error',
    danger = 'danger',
    white = 'white',
    grey = 'grey',
    black = 'black',
    info = 'info',
    graph1 = 'graph-1',
    graph2 = 'graph-2',
    graph3 = 'graph-3',
    graph4 = 'graph-4',
    graph5 = 'graph-5',
    graph6 = 'graph-6',
    graph7 = 'graph-7',
    graph8 = 'graph-8',
    graph9 = 'graph-9',
    graph10 = 'graph-10',
}